backend/requirements.txt
fastapi
uvicorn
openai
feedparser
requests
beautifulsoup4
python-dotenv
pydantic
apscheduler
